Set default port number values for the unit tests
The way in which the tests are run has been modified. A script is
now invoked when 'npm test' is run from the command line. The
script sets default values for the port numbers (PORT & PORT1 use
8080, PORT2 uses 8081).

For now, if port numbers other than the defaults are needed you'll
have to manually edit ./tests/runTests.js. When time permits we'll
change this.